+++ doc/changelog.rst
@@ -5,6 +5,7 @@ Version 0.8.1
 -------------
 
 * Rename ``metadata`` command to ``collect_metadata``
+* Add a new ``metadata`` command
 * timeit: add ``--duplicate`` option to reduce the overhead of the outer loop.
 * BenchmarkSuite constructor now requires a non-empty sequence of Benchmark
   objects.