+++ doc/changelog.rst
@@ -10,8 +10,8 @@ Version 0.8.1
 * timeit: add ``--duplicate`` option to reduce the overhead of the outer loop.
 * BenchmarkSuite constructor now requires a non-empty sequence of Benchmark
   objects.
-* metadata: store date with microsecond resolution.
-* metadata: add ``--output`` command line option.
+* Store date in metadata with microsecond resolution.
+* ``collect_metadata``: add ``--output`` command line option.
 
 Version 0.8.0 (2016-10-14)
 --------------------------